update method fix with following fixes

+++ lecture5-hw1/src/business/concretes/UserManager.java
@@ -2,6 +2,7 @@ package business.concretes;
 
 import business.abstracts.UserCheckService;
 import business.abstracts.UserService;
+import core.abstracts.EmailService;
 import dataAccess.abstracts.UserDao;
 import entities.concretes.User;
 
@@ -9,15 +10,23 @@ public class UserManager implements UserService {
 
     private final UserDao userDao;
     private final UserCheckService userCheckService;
+    private final EmailService emailService;
 
-    public UserManager(UserDao userDao, UserCheckService userCheckService) {
+    public UserManager(UserDao userDao, UserCheckService userCheckService, EmailService emailService) {
         this.userDao = userDao;
         this.userCheckService = userCheckService;
+        this.emailService = emailService;
     }
 
     @Override
     public void register(User user) {
-        if (userCheckService.isValidUser(user, userDao)) {
+        if (userCheckService.tryAuthService()) {
+            userDao.add(user);
+            System.out.println("Registration Successful");
+            return;
+        } else if ((userCheckService.isValidUser(user, userDao) && !userCheckService.isUsedEmail(user.getEmail(), userDao))) {
+            emailService.sendVerificationMail(user.getEmail());
+            System.out.println("Verified✓");
             userDao.add(user);
             System.out.println("Registration Successful");
             return;
@@ -27,7 +36,7 @@ public class UserManager implements UserService {
 
     @Override
     public void login(String email, String password) {
-        if (userCheckService.isCorrectLoginInput(email, password, userDao)) {
+        if (userCheckService.isValidLogin(email, password, userDao)) {
             System.out.println("User logged in...");
             return;
         }
@@ -36,8 +45,20 @@ public class UserManager implements UserService {
 
     @Override
     public void update(User user) {
-        if (userCheckService.isValidUser(user, userDao)) {
-            userDao.update(user);
+        User oldUser = userDao.getById(user.getId());
+        if (oldUser == null) {
+            System.out.println("Update Failed : User id does not match any of user in our database!");
+            return;
+        } else if(userCheckService.tryAuthService()){
+            System.out.println("Because you approved by google, you cannot change your infos for now :(");
+        } else if (!userCheckService.isThereAnyChange(user,oldUser)){
+            System.out.println("No change detected!");
+        } else if (userCheckService.isValidUser(user, userDao)) {
+            if (!oldUser.getEmail().equals(user.getEmail())) {
+                emailService.sendVerificationMail(user.getEmail());
+                System.out.println("Verified✓");
+            }
+            userDao.update(user, oldUser);
             System.out.println("Update Successful");
             return;
         }
@@ -46,7 +67,13 @@ public class UserManager implements UserService {
 
     @Override
     public void delete(User user) {
-        userDao.delete(user);
-        System.out.println("Delete Successful");
+        if (userDao.getById(user.getId()) != null) {
+            userDao.delete(user);
+            System.out.println("Delete Successful");
+            return;
+        }
+        System.out.println("Delete Failed: There is no such a user...");
     }
+
+
 }
